# colexec: improve the ordered synchronizer a bit