The ordered synchronizer maintains a min heap among the first rows from
each of its inputs. Whenever the current minimum row is put into the
output, the corresponding input is advanced by one row right away. In
other words, we might be eagerly fetching the next batch from that input
even if no more rows are needed.

This commit changes the behavior so that the advancement occurs only
when a new row is needed. In particular, this could reduce the amount of
rows fetched when the query has a limit.

Release note: None

Release justification: low risk, high benefit changes to existing
functionality.