@yuzefovich 2022-03-14
TFTR!

> Did any benchmarks change as a result?

I didn't check, but I highly doubt it (because our benchmarks exhaust the inputs). I guess your point maybe about the regressions - I'll run the microbenchmarks right now, before merging.

This came up when I was working on a fix for a customer in #77706 (where I had to use the limit hint twice with a hack because of this eager fetching in the ordered synchronizer).