@yuzefovich 2022-03-14
Hm, the microbenchmark says there is a minor improvement 🤷 :
```
name                    old time/op    new time/op    delta
OrderedSynchronizer-24    82.0µs ±11%    78.4µs ± 1%  -4.40%  (p=0.019 n=10+10)

name                    old speed      new speed      delta
OrderedSynchronizer-24   301MB/s ±11%   314MB/s ± 1%  +4.30%  (p=0.019 n=10+10)

name                    old alloc/op   new alloc/op   delta
OrderedSynchronizer-24     2.00B ± 0%     2.00B ± 0%    ~     (all equal)

name                    old allocs/op  new allocs/op  delta
OrderedSynchronizer-24      0.00           0.00         ~     (all equal)
```

bors r+