@yuzefovich 2022-03-14
I think whatever fix we make for #77715, we'll want it to be backported to 22.1 and 21.2, so I think it's worthwhile to backport this change to 21.2 as well.

blathers backport 21.2